Rework payment item view
