+++ Sources/ProcessOutUI/Sources/Modules/DynamicCheckout/View/Items/DynamicCheckoutPaymentItemView.swift
@@ -1,0 +1,72 @@
+//
+//  DynamicCheckoutPaymentItemView.swift
+//  ProcessOutUI
+//
+//  Created by Andrii Vysotskyi on 22.03.2024.
+//
+
+import SwiftUI
+@_spi(PO) import ProcessOutCoreUI
+
+@available(iOS 14, *)
+struct DynamicCheckoutPaymentItemView: View {
+
+    let item: DynamicCheckoutViewModelItem.Payment
+
+    var body: some View {
+        VStack {
+            HStack {
+                POAsyncImage(resource: item.iconImageResource) {
+                    Color(item.brandColor).frame(width: 24, height: 24)
+                }
+                Text(item.title)
+                    .lineLimit(1)
+                Spacer()
+                Button(
+                    action: {
+                        item.isSelected = true
+                    },
+                    label: {
+                        EmptyView()
+                    }
+                )
+                .buttonStyle(.radio) // todo(andrii-vysotskyi): fix style
+                .radioButtonSelected(item.isSelected)
+            }
+            if let information = item.additionalInformation {
+                body(information: information)
+            }
+        }
+        .contentShape(.rect)
+        .onTapGesture {
+            item.isSelected = true
+        }
+        .backport.geometryGroup()
+    }
+
+    // MARK: - Private Properties
+
+    @Environment(\.layoutDirection)
+    private var layoutDirection
+
+    @Environment(\.dynamicCheckoutStyle.paymentHeader)
+    private var style
+
+    // MARK: - Private Methods
+
+    @ViewBuilder
+    private func body(information: String) -> some View {
+        Label(
+            title: {
+                Text(information)
+                    .textStyle(style.informationText)
+            },
+            icon: {
+                Image(.info)
+                    .renderingMode(.template)
+                    .foregroundColor(style.informationText.color)
+            }
+        )
+        .frame(maxWidth: .infinity, alignment: .leading)
+    }
+}
